pyproject.toml
[project]
name = "finzup-api"
version = "0.1.0"
description = "Invoice processing API"
authors = [
    {name = "Adiel", email = "your.email@example.com"}
]
dependencies = [
    "python-multipart>=0.0.6",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pydantic[email]>=2.6.0",
    "pydantic-settings>=2.1.0",
    "langchain>=0.1.0",
    "langchain-google-genai>=0.0.5",
    "google-generativeai>=0.3.2",
    "supabase>=2.3.0",
    "python-dotenv>=1.0.0",
    "pytest>=8.0.0",
    "httpx>=0.26.0",
    "pytest-asyncio>=0.23.5",
    "pytest-cov>=4.1.0",
    "pillow>=11.2.1",
    "PyMuPDF>=1.23.8",
    "uvicorn>=0.34.2",
    "fastapi>=0.115.12",
]
requires-python = ">=3.12"
readme = "README.md"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
